backend/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.2.1
django-bootstrap5==23.1
django-phonenumber-field==7.1.0
fonttools==4.39.4
html5lib==1.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pdf2image==1.16.3
phonenumbers==8.13.11
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
PyPDF2==3.0.1
pyphen==0.14.0
pytesseract==0.3.10
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
